backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AWS Services
boto3==1.34.34
botocore==1.34.34

# LangChain for RAG
langchain==0.1.4
langchain-community==0.0.16

# Vector Store
faiss-cpu==1.13.2

# Embeddings
sentence-transformers==2.3.1
torch==2.5.1

# Image Processing
Pillow==10.2.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Groq API
groq==0.4.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis (optional)
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3

# Utilities
python-dotenv==1.0.0
pydantic-core==2.14.6
